Lifted state from Tamagotchi to App
