+++ src/components/areas/areas.tsx
@@ -1,5 +1,5 @@
 import { useState, useContext, useCallback } from "react";
-import { pokemonDataContext, pokemonHappinessContext } from "../../Tamagotchi";
+import { pokemonDataContext, pokemonHappinessContext } from "../../App";
 import {
   FeedButton,
   GetNewPokemonButton,
